requirements.txt
numpy==1.16.6
scipy==1.2.3
scikit-learn==0.20.4
scikit-MDR==0.4.4
pandas==0.24.2
joblib==0.14.1
deap==1.3.1
update-checker==0.16
tqdm==4.24.0
stopit==1.1.2
TPOT==0.9.3
networkx>=2.2
seaborn==0.8.1
matplotlib==2.2.5
pyyaml>=5.3.1
xgboost